requirements.txt
fastmcp
risingwave-py
asyncio
python-dotenv